+++ shesha-reactjs/src/designer-components/settingsInput/utils.tsx
@@ -8,3 +8,15 @@ export const getValueFromString = (string: string) => {
             return [];
     }
 };
+
+export const getWidth = (type: string, width: any) => {
+    switch (type) {
+        case 'numberField': return width || 100;
+        case 'button': return width || 24;
+        case 'dropdown': return width || 120;
+        case 'radio': return width || 120;
+        case 'colorPicker': return width || 24;
+        case 'customDropdown': return width || 120;
+        default: return width || 50;
+    }
+};
