+++ shesha-reactjs/src/designer-components/settingsInputRow/index.tsx
@@ -2,11 +2,41 @@ import React from 'react';
 
 import { IConfigurableFormComponent, IToolboxComponent } from "@/interfaces";
 import { SettingOutlined } from "@ant-design/icons";
-import { IInputRowProps, InputRow } from '../inputComponent';
+import { IInputRowProps } from '../inputComponent';
+import { SettingInput } from '../settingsInput/settingsInput';
+import { getWidth } from '../settingsInput/utils';
+import { evaluateString, useFormData } from '@/index';
+import { useStyles } from '../inputComponent/styles';
 
 export interface ISettingsInputRowProps extends Omit<IConfigurableFormComponent, 'label' | 'layout' | 'readOnly' | 'style' | 'propertyName'>, IInputRowProps {
 
 }
+
+
+export const InputRow: React.FC<IInputRowProps> = ({ inputs, readOnly, children, inline, hidden }) => {
+    const { styles } = useStyles();
+    const { data: formData } = useFormData();
+
+    const isHidden = typeof hidden === 'string' ? evaluateString(hidden, { data: formData }) : hidden;
+    return isHidden || inputs.length === 0 ? null : <div className={inline ? styles.inlineInputs : styles.rowInputs}>
+        {inputs.map((props, i) => {
+            const { type } = props;
+            const isHidden = typeof props.hidden === 'string' ? evaluateString(props.hidden, { data: formData }) : props.hidden;
+
+            const width = getWidth(type, props.width);
+
+            return (
+                <SettingInput key={i + props.label}
+                    {...props}
+                    hidden={isHidden as boolean}
+                    readOnly={readOnly}
+                    inline={inline}
+                    width={width} />
+            );
+        })}
+        {children}
+    </div>;
+};
 
 const SettingsInputRow: IToolboxComponent<ISettingsInputRowProps> = {
     type: 'settingsInputRow',
